Rearrange code generation into platform specific modules.
