+++ crates/ubrn_cli/src/codegen/mod.rs
@@ -11,7 +11,7 @@ use std::{cell::OnceCell, collections::BTreeMap, rc::Rc};
 use ubrn_bindgen::ModuleMetadata;
 use ubrn_common::{mk_dir, CrateMetadata};
 
-use crate::{config::ProjectConfig, Platform};
+use crate::config::ProjectConfig;
 
 pub(crate) trait RenderedFile: DynTemplate {
     fn path(&self, project_root: &Utf8Path) -> Utf8PathBuf;
@@ -27,18 +27,8 @@ pub(crate) trait RenderedFile: DynTemplate {
             .expect("Expected this file to have a directory");
         pathdiff::diff_utf8_paths(to, from).expect("Should be able to find a relative path")
     }
-    fn platform(&self) -> Option<Platform> {
-        None
-    }
-    fn filter_by(&self, platform: &Option<Platform>) -> bool {
-        platforms_match(platform, &self.platform())
-    }
-}
-
-fn platforms_match(platform: &Option<Platform>, this_file_wants: &Option<Platform>) -> bool {
-    match (platform, this_file_wants) {
-        (Some(building_for), Some(this_file)) => building_for == this_file,
-        (_, _) => true,
+    fn filter_by(&self) -> bool {
+        true
     }
 }
 
@@ -66,16 +56,19 @@ impl TemplateConfig {
     }
 }
 
-pub(crate) fn render_files(
-    platform: Option<Platform>,
+pub(crate) fn get_template_config(
     project: ProjectConfig,
     rust_crate: CrateMetadata,
     modules: Vec<ModuleMetadata>,
-) -> Result<()> {
-    let config = Rc::new(TemplateConfig::new(project, rust_crate, modules));
-    let files = files::get_files(config.clone());
-    let files = files.iter().filter(|f| f.filter_by(&platform)).cloned();
+) -> Rc<TemplateConfig> {
+    Rc::new(TemplateConfig::new(project, rust_crate, modules))
+}
 
+pub(crate) fn render_files(
+    config: Rc<TemplateConfig>,
+    files: impl Iterator<Item = Rc<dyn RenderedFile>>,
+) -> Result<()> {
+    let files = files.filter(|f| f.filter_by());
     let project_root = config.project.project_root();
     let map = render_templates(project_root, files)?;
     let exclude_files = config.project.exclude_files();
@@ -109,6 +102,7 @@ fn render_templates(
     Ok(map)
 }
 
+#[macro_export]
 macro_rules! templated_file {
     ($T:ty, $filename:literal) => {
         paste::paste! {
@@ -132,294 +126,39 @@ macro_rules! templated_file {
     };
 }
 
-mod files {
-    use super::platforms_match;
-    use super::RenderedFile;
-    use super::TemplateConfig;
-    use crate::Platform;
-    use camino::Utf8Path;
-    use camino::Utf8PathBuf;
+pub(crate) mod files {
     use std::rc::Rc;
 
-    impl TemplateConfig {
-        fn uses_kotlin(self: &Rc<Self>) -> bool {
-            *self.uses_kotlin.get_or_init(|| {
-                let project_root = self.project.project_root();
-                let gradle_file = BuildGradle::new(self.clone()).path(project_root);
-                if gradle_file.exists() {
-                    let file = std::fs::read_to_string(gradle_file)
-                        .expect("Cannot read build.gradle file");
-                    file.contains("kotlin")
-                } else {
-                    // assume that if the user blew away the gradle file,
-                    // then we should remake it as one with kotlin.
-                    true
-                }
-            })
+    use super::{RenderedFile, TemplateConfig};
+    use crate::{jsi, wasm, Platform};
+
+    pub(crate) fn get_files_for(
+        config: Rc<TemplateConfig>,
+        platform: &Platform,
+    ) -> Vec<Rc<dyn RenderedFile>> {
+        let mut files = vec![];
+        match platform {
+            Platform::Android => {
+                files.extend(jsi::crossplatform::get_files(config.clone()));
+                files.extend(jsi::android::get_files(config.clone()));
+            }
+            Platform::Ios => {
+                files.extend(jsi::crossplatform::get_files(config.clone()));
+                files.extend(jsi::ios::get_files(config.clone()));
+            }
+            #[cfg(feature = "wasm")]
+            Platform::Wasm => {
+                files.extend(wasm::get_files(config.clone()));
+            }
         }
+        files
     }
 
-    // The set of files to be generated. These are filtered depending on what platform we're
-    // building for, and based upon the template config.
-    pub(super) fn get_files(config: Rc<TemplateConfig>) -> Vec<Rc<dyn RenderedFile>> {
-        vec![
-            // typescript
-            IndexTsx::rc_new(config.clone()),
-            // C++
-            TMHeader::rc_new(config.clone()),
-            TMCpp::rc_new(config.clone()),
-            // Codegen (for installer)
-            NativeCodegenTs::rc_new(config.clone()),
-            // Android
-            CMakeLists::rc_new(config.clone()),
-            CppAdapter::rc_new(config.clone()),
-            AndroidManifest::rc_new(config.clone()),
-            // Android with Java
-            JavaModule::rc_new(config.clone()),
-            JavaPackage::rc_new(config.clone()),
-            BuildGradle::rc_new(config.clone()),
-            // Android with Kotlin
-            KtModule::rc_new(config.clone()),
-            KtPackage::rc_new(config.clone()),
-            KtBuildGradle::rc_new(config.clone()),
-            // iOS
-            ModuleTemplateH::rc_new(config.clone()),
-            ModuleTemplateMm::rc_new(config.clone()),
-            PodspecTemplate::rc_new(config.clone()),
-        ]
-    }
-
-    templated_file!(JavaModule, "ModuleTemplate.java");
-    impl RenderedFile for JavaModule {
-        fn path(&self, project_root: &Utf8Path) -> Utf8PathBuf {
-            let name = self.config.project.module_cpp();
-            let filename = format!("{name}Module.java");
-            self.config
-                .project
-                .android
-                .codegen_package_dir(project_root)
-                .join(filename)
-        }
-        fn filter_by(&self, platform: &Option<Platform>) -> bool {
-            platforms_match(platform, &self.platform()) && !self.config.uses_kotlin()
-        }
-        fn platform(&self) -> Option<Platform> {
-            Some(Platform::Android)
-        }
-    }
-
-    templated_file!(JavaPackage, "PackageTemplate.java");
-    impl RenderedFile for JavaPackage {
-        fn path(&self, project_root: &Utf8Path) -> Utf8PathBuf {
-            let name = self.config.project.module_cpp();
-            let filename = format!("{name}Package.java");
-            self.config
-                .project
-                .android
-                .codegen_package_dir(project_root)
-                .join(filename)
-        }
-        fn filter_by(&self, platform: &Option<Platform>) -> bool {
-            platforms_match(platform, &self.platform()) && !self.config.uses_kotlin()
-        }
-        fn platform(&self) -> Option<Platform> {
-            Some(Platform::Android)
-        }
-    }
-
-    templated_file!(BuildGradle, "build.gradle");
-    impl RenderedFile for BuildGradle {
-        fn path(&self, project_root: &Utf8Path) -> Utf8PathBuf {
-            let filename = "build.gradle";
-            self.config
-                .project
-                .android
-                .directory(project_root)
-                .join(filename)
-        }
-        fn filter_by(&self, platform: &Option<Platform>) -> bool {
-            platforms_match(platform, &self.platform()) && !self.config.uses_kotlin()
-        }
-        fn platform(&self) -> Option<Platform> {
-            Some(Platform::Android)
-        }
-    }
-
-    templated_file!(KtModule, "ModuleTemplate.kt");
-    impl RenderedFile for KtModule {
-        fn path(&self, project_root: &Utf8Path) -> Utf8PathBuf {
-            let name = self.config.project.module_cpp();
-            let filename = format!("{name}Module.kt");
-            self.config
-                .project
-                .android
-                .codegen_package_dir(project_root)
-                .join(filename)
-        }
-        fn filter_by(&self, platform: &Option<Platform>) -> bool {
-            platforms_match(platform, &self.platform()) && self.config.uses_kotlin()
-        }
-        fn platform(&self) -> Option<Platform> {
-            Some(Platform::Android)
-        }
-    }
-
-    templated_file!(KtPackage, "PackageTemplate.kt");
-    impl RenderedFile for KtPackage {
-        fn path(&self, project_root: &Utf8Path) -> Utf8PathBuf {
-            let name = self.config.project.module_cpp();
-            let filename = format!("{name}Package.kt");
-            self.config
-                .project
-                .android
-                .codegen_package_dir(project_root)
-                .join(filename)
-        }
-        fn filter_by(&self, platform: &Option<Platform>) -> bool {
-            platforms_match(platform, &self.platform()) && self.config.uses_kotlin()
-        }
-        fn platform(&self) -> Option<Platform> {
-            Some(Platform::Android)
-        }
-    }
-
-    templated_file!(KtBuildGradle, "build.kt.gradle");
-    impl RenderedFile for KtBuildGradle {
-        fn path(&self, project_root: &Utf8Path) -> Utf8PathBuf {
-            let filename = "build.gradle";
-            self.config
-                .project
-                .android
-                .directory(project_root)
-                .join(filename)
-        }
-        fn filter_by(&self, platform: &Option<Platform>) -> bool {
-            platforms_match(platform, &self.platform()) && self.config.uses_kotlin()
-        }
-        fn platform(&self) -> Option<Platform> {
-            Some(Platform::Android)
-        }
-    }
-
-    templated_file!(CMakeLists, "CMakeLists.txt");
-    impl RenderedFile for CMakeLists {
-        fn path(&self, project_root: &Utf8Path) -> Utf8PathBuf {
-            let filename = "CMakeLists.txt";
-            self.config
-                .project
-                .android
-                .directory(project_root)
-                .join(filename)
-        }
-        fn platform(&self) -> Option<Platform> {
-            Some(Platform::Android)
-        }
-    }
-
-    templated_file!(TMHeader, "TurboModuleTemplate.h");
-    impl RenderedFile for TMHeader {
-        fn path(&self, project_root: &Utf8Path) -> Utf8PathBuf {
-            let filename = format!("{}.h", self.config.project.cpp_filename());
-            self.config.project.tm.cpp_path(project_root).join(filename)
-        }
-    }
-
-    templated_file!(TMCpp, "TurboModuleTemplate.cpp");
-    impl RenderedFile for TMCpp {
-        fn path(&self, project_root: &Utf8Path) -> Utf8PathBuf {
-            let filename = format!("{}.cpp", self.config.project.cpp_filename());
-            self.config.project.tm.cpp_path(project_root).join(filename)
-        }
-    }
-
-    templated_file!(IndexTsx, "index.tsx");
-    impl RenderedFile for IndexTsx {
-        fn path(&self, project_root: &Utf8Path) -> Utf8PathBuf {
-            let filename = "index.tsx";
-            self.config.project.tm.ts_path(project_root).join(filename)
-        }
-    }
-
-    templated_file!(NativeCodegenTs, "NativeCodegenTemplate.ts");
-    impl RenderedFile for NativeCodegenTs {
-        fn path(&self, project_root: &Utf8Path) -> Utf8PathBuf {
-            let filename = format!("{}.ts", self.config.project.codegen_filename());
-            self.config.project.tm.ts_path(project_root).join(filename)
-        }
-    }
-
-    templated_file!(CppAdapter, "cpp-adapter.cpp");
-    impl RenderedFile for CppAdapter {
-        fn path(&self, project_root: &Utf8Path) -> Utf8PathBuf {
-            let filename = "cpp-adapter.cpp";
-            self.config
-                .project
-                .android
-                .directory(project_root)
-                .join(filename)
-        }
-        fn platform(&self) -> Option<Platform> {
-            Some(Platform::Android)
-        }
-    }
-
-    templated_file!(ModuleTemplateH, "ModuleTemplate.h");
-    impl RenderedFile for ModuleTemplateH {
-        fn path(&self, project_root: &Utf8Path) -> Utf8PathBuf {
-            let name = self.config.project.module_cpp();
-            let filename = format!("{name}.h");
-            self.config
-                .project
-                .ios
-                .directory(project_root)
-                .join(filename)
-        }
-        fn platform(&self) -> Option<Platform> {
-            Some(Platform::Ios)
-        }
-    }
-
-    templated_file!(ModuleTemplateMm, "ModuleTemplate.mm");
-    impl RenderedFile for ModuleTemplateMm {
-        fn path(&self, project_root: &Utf8Path) -> Utf8PathBuf {
-            let name = self.config.project.module_cpp();
-            let filename = format!("{name}.mm");
-            self.config
-                .project
-                .ios
-                .directory(project_root)
-                .join(filename)
-        }
-        fn platform(&self) -> Option<Platform> {
-            Some(Platform::Ios)
-        }
-    }
-
-    templated_file!(PodspecTemplate, "module-template.podspec");
-    impl RenderedFile for PodspecTemplate {
-        fn path(&self, project_root: &Utf8Path) -> Utf8PathBuf {
-            let name = self.config.project.podspec_filename();
-            let filename = format!("{name}.podspec");
-            project_root.join(filename)
-        }
-        fn platform(&self) -> Option<Platform> {
-            Some(Platform::Ios)
-        }
-    }
-
-    templated_file!(AndroidManifest, "AndroidManifest.xml");
-    impl RenderedFile for AndroidManifest {
-        fn path(&self, project_root: &Utf8Path) -> Utf8PathBuf {
-            let filename = "AndroidManifest.xml";
-            self.config
-                .project
-                .android
-                .src_main_dir(project_root)
-                .join(filename)
-        }
-        fn platform(&self) -> Option<Platform> {
-            Some(Platform::Android)
-        }
+    pub(crate) fn get_files(config: Rc<TemplateConfig>) -> Vec<Rc<dyn RenderedFile>> {
+        let mut files = vec![];
+        files.extend(jsi::get_files(config.clone()));
+        files.extend(wasm::get_files(config.clone()));
+        files
     }
 }
 
@@ -530,23 +269,5 @@ mod tests {
         assert!(s.contains("module_cpp = MyTesterTemplateProject."));
         assert!(s.contains("list of modules = ['NativeAlice', 'NativeBob', 'NativeCharlie']"));
         Ok(())
-    }
-
-    #[test]
-    fn test_platform_matching() {
-        let not_specfied: Option<Platform> = None;
-        let android_only = Some(Platform::Android);
-        let ios_only = Some(Platform::Ios);
-        assert!(platforms_match(&not_specfied, &not_specfied));
-        assert!(platforms_match(&not_specfied, &android_only));
-        assert!(platforms_match(&not_specfied, &ios_only));
-
-        assert!(platforms_match(&android_only, &not_specfied));
-        assert!(platforms_match(&android_only, &android_only));
-        assert!(!platforms_match(&android_only, &ios_only));
-
-        assert!(platforms_match(&ios_only, &not_specfied));
-        assert!(!platforms_match(&ios_only, &android_only));
-        assert!(platforms_match(&ios_only, &ios_only));
     }
 }
