+++ crates/ubrn_cli/src/wasm/mod.rs
@@ -3,6 +3,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/
  */
+mod codegen;
 mod generate;
 
+pub(crate) use codegen::get_files;
 pub(crate) use generate::CmdArg;
