+++ crates/ubrn_cli/templates/jsi/android/AndroidManifest.xml
