+++ crates/ubrn_cli/templates/jsi/android/CMakeLists.txt
