+++ crates/ubrn_cli/templates/jsi/android/ModuleTemplate.kt
