+++ crates/ubrn_cli/templates/jsi/android/PackageTemplate.kt
