+++ crates/ubrn_cli/templates/jsi/android/cpp-adapter.cpp
