+++ crates/ubrn_cli/templates/jsi/crossplatform/NativeCodegenTemplate.ts
