+++ crates/ubrn_cli/templates/jsi/ios/module-template.podspec
